Start working on updated C backend with more flexible options
Options can be read from a JSON config file, which avoids having to list
lots of options on the command line, and it lets us include things like
lists of definitions to export excetera. This is done in a backwards
compatible way, but currently I have it in c_backend2.ml for testing
right now.

The next step is to tease apart the final compile_ast function to
generate various files according to the config.